Organize the code (#13)
