+++ eCommerce/eCommerce.csproj
@@ -8,4 +8,9 @@
       <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.10.0" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Adapters" />
+      <Folder Include="Auth" />
+    </ItemGroup>
+
 </Project>
